update highscores.js, add highscoresState.js and highscoresTable.js
