--- src/highscores/highscoresInvokeFunction.js
@@ -1,3 +1,0 @@
-import { displayHighScores } from "./highscores.js";
-
-displayHighScores();
